# Improve performance of buildtsi tool / series creation