Fixes #10019 

This PR significantly improves the performance of `influx_inspect buildtsi`. It makes building a TSI index around 4-10 times faster than 1.5.3, and 17–46 times faster than current `master`. 😃 

#### New tool features

 - Shards will be processed concurrently (adjustable with `-c`);
 - Series batch size configurable. This is an "expert" option, and users should typically stick with the default value, which is 10,000; and
 - Much. More. Fasterer.

#### Test data

I wrote 100,000,000 series with inch across 25 shards.

  - 4 measurements
  - 5 tag keys

```
inch -b 10000 -c 5 -t 10,10,10,100,1000 -f 1 -m 4 -p 1 -start-time '2018-07-02T00:00:00Z' -time 100h  -v
```

#### TLDR Results

These results show the time taken to build the TSI index with a pre-existing series file. Most of the performance boost comes from changes to how the index is built so I focussed on that results-wise:

| Description | master time | PR time| time Δ  | master series/s | PR series/s | series/s Δ |
| ---: | :---: | :---: | :---: | :---: | :---: | :---: |
| single shard (4M series) |  8m 15s | 28s | ~94% |  9K  | ~143K  | ~94% |
| 25 shards (100M series) |  ~200 min | 4m 20s | ~98% |  ~8.3K  | ~﻿384K  | ~98% |

N.B these results were against `master`. 1.5.3 might not be _as bad_ as the master results because it doesn't contain #9982, which hurts `buildtsi` performance. 

#### Breakdown

There are some other useful commits in the set, but these are the ones that made the most difference:

 - f6facb5 and 2c40e80: we have to flush buffers and syncing files on every write to a `LogFile` for durability reasons. However, we don't need to do this for offline tooling since in the worst-case failure, we can just re-run the tool. Let's turn that off! 
 - 10a45f3: rather than reading series one at a time out of a TSM file, we can read a batch and then send them all at once to be written to the index. 10,000 is a good, empirically determined, default batch size.
 - f78ff85: A `LogEntry` is 11-12 bytes per series. If you write a batch of 10,000 new series that's going to involve three fsyncs. By making the buffer configurable for the `buildtsi` tool we can reduce the frequency of these syscalls.
 - 4792ba4: We maintain inverted indexes for `measurement_name -> all_series_with_name` amd `tag_key_value -> all_series_with_tag_key_value`. Rather than storing the sets of series ids for these indexes in maps, we use bitmaps. This reduces their size and improves performance when we need to pull them out in order. This will significantly help with TSI query planning too. Note: bitsets have a non-trivial cost for low cardinality sets (`~1000 ns` and `~15` allocs),  so we don't always use them.
 - 0bf0180: re-use tag containers... This one I need some scrutiny on. I'm almost certain that it is OK to do this for the offline tooling, but please take a close look @benbjohnson @stuartcarnie.... It made a massive improvement: `~37s -> ~28s`.
 - a85e951 concurrent shard index building. This will help for machines with many cores. How it helps with machines with fewer than 8 OS-threads is unclear because the series file and index already use multiple cores during insertion anyway.


#### Todo 

 I want to abstract a bunch of the "when to use bitsets and when not to" into the `tsdb.SeriesIDSet`. I will do that in a future PR.