@e-dard 2018-07-05
Nit: I forgot to set a limit of 128K on the log buffer size in f78ff85. I will add a commit to the set to do that.

Incidentally, 128K seems to work pretty well for me, but I'm not sure if it's maybe at bit high?