@e-dard 2018-07-09
Final performance improvements on a box with 24 OS threads:

| Description | v1.5.3 time | PR time| time Δ 
| ---: | :---: | :---: | :---: | 
| single shard (4M series) |  2m 2s | 12s | ~90% | 
| 25 shards (100M series) |  ~51m 40s | 1m 15s | ~97.5% |